# docs/api.openapi.yaml
openapi: 3.1.0
info:
  title: Evernest API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Baby tracking backend. All requests/responses are JSON. All authenticated
    endpoints require `Authorization: Bearer <access_token>`. The refresh
    token is returned in the JSON body of `/v1/auth/{register,login,refresh}`
    and the client persists it itself (e.g. `localStorage`); it is passed
    back in the body of `/v1/auth/{refresh,logout}`. The legacy
    `evernest_refresh` `httpOnly` cookie is still set and still accepted as
    a fallback during a migration window — see the `RefreshRequest` and
    `refreshCookie` schemas — but new clients should ignore the cookie.

    ## Idempotency

    Every event POST (bottle-feeds, diapers, pumpings, nursing-sessions,
    growths, sleeps) accepts a client-supplied `id` (UUIDv7) in the request body.
    Inserts use `ON CONFLICT (id) DO NOTHING`, so retrying a write with the
    same `id` returns the existing row with status `201` instead of
    erroring. If `id` is omitted the server generates one. There is no
    separate `Idempotency-Key` header — the body field IS the idempotency
    key.

    ## Errors

    All non-2xx responses share the same envelope:

    ```json
    { "error": { "code": "validation_failed", "message": "..." } }
    ```

    `code` is a stable machine-readable string (e.g. `validation_failed`,
    `unauthorized`, `not_found`). `message` is human-readable.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/babies/{babyID}/sleeps:
    parameters:
      - $ref: "#/components/parameters/BabyIDPath"
    post:
      tags: [Sleeps]
      operationId: createSleep
      summary: Log a sleep session
      description: |
        Idempotent on the request body's `id` field. Omitting `ended_at`
        opens an in-progress session ("baby just fell asleep"); at most
        one open session per baby is allowed. `ended_at`, when provided,
        must be `>= started_at`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSleepRequest"
      responses:
        "201":
          description: Sleep created (or returned unchanged on idempotent replay).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sleep"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "`open_session_exists` — a sleep is already in progress for this baby."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Sleeps]
      operationId: listSleeps
      summary: List sleep sessions for a baby
      parameters:
        - $ref: "#/components/parameters/FromQuery"
        - $ref: "#/components/parameters/ToQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Sleeps in the [from, to) window (by `started_at`), newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sleep"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/sleeps/{id}:
    parameters:
      - $ref: "#/components/parameters/IDPath"
    patch:
      tags: [Sleeps]
      operationId: updateSleep
      summary: Close an open sleep OR edit a closed one
      description: |
        Dispatches based on the row's current `ended_at` state:

        - **Open session** (`ended_at IS NULL`): closes the session.
          Body must include `ended_at` (`>= started_at`). Returns
          `409 already_closed` if the row was closed by a concurrent
          request.
        - **Closed session** (`ended_at IS NOT NULL`): partial edit.
          All fields optional. Sending `ended_at: null` is not
          supported (edits can't re-open a closed session). Set
          `clear_sleep_type: true` to clear the type; send an empty
          string for `location` or `notes` to clear them. Returns
          `409 session_state_changed` if the row was deleted or
          re-opened mid-edit.

        Server preserves `id`, `source`, `created_at`, and
        `created_by_user_id` regardless of input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/EndSleepRequest"
                - $ref: "#/components/schemas/UpdateSleepRequest"
      responses:
        "200":
          description: Closed (open session path) or updated (closed session path).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sleep"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Either `already_closed` (close-path) or
            `session_state_changed` (edit-path) — see description.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Sleeps]
      operationId: deleteSleep
      summary: Delete a sleep session
      description: Idempotent. Open rows can be deleted ("started by mistake").
      responses:
        "204":
          description: Deleted (or no row matched).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

  # -----------------------------------------------------------------------
  # Coming soon
  # -----------------------------------------------------------------------
  # CP3 Charts will add `GET /v1/babies/{babyID}/charts/daily` in its own
  # PR. The schema for that endpoint is intentionally NOT defined here —
  # see the parallel CP3 Charts agent's PR for the full spec.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HMAC-signed JWT access token issued by `/v1/auth/{register,login,refresh}`.
        Send as `Authorization: Bearer <token>`.
    refreshCookie:
      type: apiKey
      in: cookie
      name: evernest_refresh
      description: |
        DEPRECATED — accepted as a fallback during a migration window so
        clients that pre-date the body-based refresh-token rollout keep
        working. New clients should ignore this cookie and use the
        `refresh_token` field in `TokenResponse` + `RefreshRequest`
        instead.

        `httpOnly`, opaque (random) refresh token. Set by
        `/v1/auth/{register,login,refresh}` and consumed (as a fallback)
        by `/v1/auth/{refresh,logout}`. Cookie path is `/v1/auth`.
...
  parameters:
...
    BabyIDPath:
      name: babyID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 01919b87-9a8c-7f2d-b1c3-4d5e6f7a8b9c
      description: Baby UUID.
    IDPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: 01919b88-1c2d-7e3f-a4b5-c6d7e8f9a0b1
      description: Resource UUID (event id).
    FromQuery:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      example: "2026-05-26T00:00:00Z"
      description: |
        Inclusive lower bound (RFC3339). Defaults to 24h ago for events,
        365 days ago for growths.
    ToQuery:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      example: "2026-05-27T00:00:00Z"
      description: |
        Exclusive upper bound (RFC3339). Defaults to now+24h. `from` must
        be strictly before `to`.
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 200
      example: 200
      description: Max rows returned. Out-of-range values are silently clamped to the default.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              example: validation_failed
            message:
              type: string
              description: Human-readable message (may be the raw validator output).
              example: "Key: 'createReq.AmountML' Error:Field validation for 'AmountML' failed on the 'gt' tag"
      example:
        error:
          code: validation_failed
          message: "Key: 'createReq.AmountML' Error:Field validation for 'AmountML' failed on the 'gt' tag"
...
    Sleep:
      type: object
      required:
        - id
        - baby_id
        - started_at
        - source
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: 01919b8c-1a2b-7c3d-e4f5-a6b7c8d9e0f1
        baby_id:
          type: string
          format: uuid
          example: 01919b87-9a8c-7f2d-b1c3-4d5e6f7a8b9c
        started_at:
          type: string
          format: date-time
          example: "2026-05-26T13:00:00Z"
        ended_at:
          type: string
          format: date-time
          description: Absent while the session is open (baby still asleep).
          example: "2026-05-26T14:25:00Z"
        sleep_type:
          type: string
          enum: [nap, night]
          example: nap
        location:
          type: string
          maxLength: 100
          example: crib
        notes:
          type: string
          maxLength: 500
          example: "fell asleep on the walk"
        source:
          type: string
          example: manual
        created_at:
          type: string
          format: date-time
          example: "2026-05-26T13:00:01Z"
    CreateSleepRequest:
      type: object
      required: [started_at]
      properties:
        id:
          type: string
          format: uuid
          description: Optional client-supplied UUIDv7 for idempotency.
          example: 01919b8c-1a2b-7c3d-e4f5-a6b7c8d9e0f1
        started_at:
          type: string
          format: date-time
          example: "2026-05-26T13:00:00Z"
        ended_at:
          type: string
          format: date-time
          description: |
            Optional. Omit to open an in-progress session; when provided,
            must be `>= started_at`.
          example: "2026-05-26T14:25:00Z"
        sleep_type:
          type: string
          enum: [nap, night]
          example: nap
        location:
          type: string
          maxLength: 100
          example: crib
        notes:
          type: string
          maxLength: 500
          example: "fell asleep on the walk"
    EndSleepRequest:
      type: object
      required: [ended_at]
      description: |
        Body shape for closing an OPEN sleep session via
        `PATCH /v1/sleeps/{id}`.
      properties:
        ended_at:
          type: string
          format: date-time
          example: "2026-05-26T14:25:00Z"
    UpdateSleepRequest:
      type: object
      description: |
        Body shape for editing a CLOSED sleep session via
        `PATCH /v1/sleeps/{id}`. All fields optional. Set
        `clear_sleep_type: true` to clear that nullable column. Send an
        empty string for `location` or `notes` to clear them.
      properties:
        started_at:
          type: string
          format: date-time
          example: "2026-05-26T13:00:00Z"
        ended_at:
          type: string
          format: date-time
          description: When provided, must be `>= started_at`.
          example: "2026-05-26T14:25:00Z"
        sleep_type:
          type: string
          enum: [nap, night]
          example: night
        clear_sleep_type:
          type: boolean
          example: false
        location:
          type: string
          maxLength: 100
          example: crib
        notes:
          type: string
          maxLength: 500
          example: "restless"
...
  responses:
    BadRequest:
      description: Malformed request body, query parameter, or path id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            badJSON:
              summary: Body could not be decoded
              value:
                error:
                  code: bad_request
                  message: "json: unknown field \"foo\""
            badID:
              summary: Path id is not a valid UUID
              value:
                error:
                  code: bad_id
                  message: invalid baby id
            badRange:
              summary: from/to query parameter is malformed or inverted
              value:
                error:
                  code: bad_range
                  message: from must be RFC3339
    Unauthorized:
      description: Missing/invalid bearer token (or, on `/v1/auth/refresh`, a missing/invalid refresh cookie).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: missing bearer token
    Forbidden:
      description: The caller is authenticated but not a member of the household that owns the resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: forbidden
              message: not a member of this household
    NotFound:
      description: The household or baby referenced in the path does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: baby not found
    ValidationError:
      description: |
        Request was syntactically valid JSON but failed semantic validation
        (struct tag rules, enum constraints, range bounds, etc.). The
        `message` field is the raw validator output.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missingField:
              summary: Required field missing
              value:
                error:
                  code: validation_failed
                  message: "Key: 'createReq.OccurredAt' Error:Field validation for 'OccurredAt' failed on the 'required' tag"
            outOfRange:
              summary: Numeric value out of range
              value:
                error:
                  code: validation_failed
                  message: "Key: 'createReq.AmountML' Error:Field validation for 'AmountML' failed on the 'gt' tag"
            growthAllNull:
              summary: Growth POST with no measurements
              value:
                error:
                  code: validation_failed
                  message: "at least one of weight_g, height_cm, head_circumference_cm is required"
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal
              message: could not create baby